Cleanup
- Removed unused packages.
- Moved modules to be more consistent.
- Cleaned up code to follow a more consistent structure.
