+++ frontend/LocalHelper/lib/Additions/Providers/authSettings.dart
