+++ frontend/LocalHelper/lib/Additions/Providers/settings.dart
