+++ frontend/LocalHelper/lib/Screens/MyPosts/screen_editposts.dart
@@ -1,36 +1,51 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:localhelper/Additions/authSettings.dart';
-import 'package:localhelper/Additions/settings.dart';
+import 'package:localhelper/Additions/Providers/authSettings.dart';
+import 'package:localhelper/Additions/Providers/settings.dart';
 import 'package:provider/provider.dart';
 import 'package:http/http.dart' as http;
 
-class ScreenCreatePosts extends StatefulWidget {
+class ScreenEditPosts extends StatefulWidget {
+  final int postId;
+  final String title;
+  final String description;
+  final bool req;
+  ScreenEditPosts(this.postId, this.title, this.description, this.req);
   @override
-  _ScreenCreatePostsState createState() => _ScreenCreatePostsState();
+  _ScreenEditPostsState createState() =>
+      _ScreenEditPostsState(title, description, req);
 }
 
-class _ScreenCreatePostsState extends State<ScreenCreatePosts> {
+class _ScreenEditPostsState extends State<ScreenEditPosts> {
+// VARIABLES ===================================================================
+
   // Text Controllers
   final titleController = TextEditingController();
-  final nameController = TextEditingController();
   final descriptionController = TextEditingController();
 
   // Prevent Multi sending
   bool enableSend = true;
-  bool request = false;
+  bool request;
+
+// =============================================================================
+// FUNCTIONS ===================================================================
+
+  // Constructor
+  _ScreenEditPostsState(String title, String description, this.request) {
+    titleController.text = title;
+    descriptionController.text = description;
+  }
 
   @override
   void dispose() {
     titleController.dispose();
-    nameController.dispose();
     descriptionController.dispose();
     super.dispose();
   }
 
-  Future<bool> sendPost(String token, String title, String desc, bool request,
-      AuthSettings authSettings) async {
+  Future<void> sendPost(
+      String token, String title, String desc, bool request) async {
     // Flutter Json
     Map<String, dynamic> jsonMap = {
       'title': title,
@@ -47,58 +62,29 @@ class _ScreenCreatePostsState extends State<ScreenCreatePosts> {
         'accept': 'application/json',
         'authorization': token,
       };
-      var response = await http.post(
-        'https://localhelper-backend.herokuapp.com/api/posts',
+      var response = await http.put(
+        'https://localhelper-backend.herokuapp.com/api/posts/' +
+            widget.postId.toString(),
         headers: headers,
         body: jsonString,
       );
-
-      // Error
-      if (response.statusCode != 201) {
-        print(response.statusCode.toString());
-        return false;
-      } else {
-        // If the zip
-        if (authSettings.zipID != -1) {
-          var postId = await http.get(
-            'https://localhelper-backend.herokuapp.com/api/posts/me',
-            headers: headers,
-          );
-
-          var postIDJson = jsonDecode(postId.body).last;
-
-          // Post the new zip
-          Map<String, dynamic> jsonMap = {
-            'postId': postIDJson['post']['id'],
-            'zipId': authSettings.zipID,
-          };
-
-          // Encode
-          String jsonString = json.encode(jsonMap);
-
-          var result = await http.post(
-            'https://localhelper-backend.herokuapp.com/api/postZips',
-            headers: headers,
-            body: jsonString,
-          );
-
-          print(result.statusCode);
-        }
-
-        return true;
+      print(response.statusCode);
+      if (response.statusCode == 200) {
+        Navigator.pop(context);
       }
     } catch (e) {
       print(e);
-      return false;
     }
   }
+
+// =============================================================================
+// MAIN ========================================================================
 
   @override
   Widget build(BuildContext context) {
     // Providers
     Settings settings = Provider.of<Settings>(context);
     AuthSettings authSettings = Provider.of<AuthSettings>(context);
-
     return GestureDetector(
       onTap: () {
         FocusScope.of(context).requestFocus(FocusNode());
@@ -115,7 +101,7 @@ class _ScreenCreatePostsState extends State<ScreenCreatePosts> {
           ),
           centerTitle: true,
           title: Text(
-            'Create New Post',
+            'Edit Post',
             style: TextStyle(
                 fontSize: 30,
                 fontWeight: FontWeight.bold,
@@ -156,10 +142,7 @@ class _ScreenCreatePostsState extends State<ScreenCreatePosts> {
             SwitchListTile(
               title: Text(
                 'Request?',
-                style: TextStyle(
-                    color: settings.darkMode ? Colors.white : Colors.black,
-                    fontSize: 30,
-                    fontWeight: FontWeight.bold),
+                style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold),
               ),
               value: request,
               onChanged: (value) {
@@ -213,27 +196,15 @@ class _ScreenCreatePostsState extends State<ScreenCreatePosts> {
                           setState(() {
                             enableSend = false;
                           });
-                          if (titleController.text.isEmpty &&
-                              descriptionController.text.isEmpty) {
-                            Navigator.pop(context, null);
-                          } else {
-                            // Send post to internet.
-                            var result = await sendPost(
-                                authSettings.token,
-                                titleController.text,
-                                descriptionController.text,
-                                request,
-                                authSettings);
-                            if (result) {
-                              settings.refreshPage();
-                              Navigator.pop(context);
-                            } else {
-                              setState(() {
-                                enableSend = true;
-                              });
-                            }
-                          }
+                          await sendPost(
+                              authSettings.token,
+                              titleController.text,
+                              descriptionController.text,
+                              request);
                         }
+                        setState(() {
+                          enableSend = true;
+                        });
                       },
                       splashColor:
                           settings.darkMode ? Colors.red : Colors.black,
@@ -242,7 +213,7 @@ class _ScreenCreatePostsState extends State<ScreenCreatePosts> {
                       minWidth: double.infinity,
                       height: 60,
                       child: Text(
-                        'Submit',
+                        'Save',
                         style: TextStyle(
                           color:
                               settings.darkMode ? Colors.white : Colors.black,
