+++ frontend/LocalHelper/lib/Screens/Settings/screen_userSettings.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
-import 'package:localhelper/Additions/authSettings.dart';
-import 'package:localhelper/Additions/settings.dart';
+import 'package:localhelper/Additions/Providers/authSettings.dart';
+import 'package:localhelper/Additions/Providers/settings.dart';
 import 'package:provider/provider.dart';
 
 class ScreenUserSettings extends StatefulWidget {
@@ -92,6 +92,8 @@ class OwnerDone extends StatefulWidget {
 }
 
 class _OwnerDoneState extends State<OwnerDone> {
+// VARIABLES ===================================================================
+
   // Save json info
   final info;
 
@@ -102,6 +104,13 @@ class _OwnerDoneState extends State<OwnerDone> {
   final phoneController = TextEditingController();
   final emailController = TextEditingController();
   final zipController = TextEditingController();
+
+  // Loading
+  bool _loading = false;
+  bool _invalidZip = false;
+
+// =============================================================================
+// FUNCTIONS ===================================================================
 
   @override
   void dispose() {
@@ -114,6 +123,7 @@ class _OwnerDoneState extends State<OwnerDone> {
     super.dispose();
   }
 
+  // Constructor
   _OwnerDoneState(this.info, AuthSettings authSettings) {
     firstNController.text = info['first'];
     lastNController.text = info['last'];
@@ -123,9 +133,246 @@ class _OwnerDoneState extends State<OwnerDone> {
     zipController.text = authSettings.zip;
   }
 
-  // Loading
-  bool _loading = false;
-  bool _invalidZip = false;
+  Future<bool> _showMyDialog() async {
+    return showDialog<bool>(
+      context: context,
+      barrierDismissible: false, // user must tap button!
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text('Delete User?'),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: <Widget>[
+                Text('This action is permanent!'),
+              ],
+            ),
+          ),
+          actions: <Widget>[
+            TextButton(
+              child: Text('No'),
+              onPressed: () {
+                Navigator.of(context).pop(false);
+              },
+            ),
+            TextButton(
+              child: Text('Yes'),
+              onPressed: () {
+                Navigator.of(context).pop(true);
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+// =============================================================================
+// WIDGETS =====================================================================
+
+// Save button function
+  Widget saveButton(Settings settings, AuthSettings authSettings) {
+    return FlatButton(
+      height: 60,
+      minWidth: double.infinity,
+      onPressed: () async {
+        if (zipController.text.length < 5) {
+          setState(() {
+            _invalidZip = true;
+          });
+          return false;
+        }
+
+        try {
+          // Put body
+          Map<String, dynamic> jsonMap = {
+            'first': firstNController.text,
+            'last': lastNController.text,
+            'gender': genderController.text,
+            'phone': phoneController.text,
+            'email': emailController.text,
+          };
+
+          // Encode
+          String jsonString = json.encode(jsonMap);
+
+          // Header
+          Map<String, String> headers = {
+            'authorization': authSettings.token,
+            "Content-Type": "application/json",
+          };
+          String link = 'https://localhelper-backend.herokuapp.com/api/users';
+          var result = await http.put(link, headers: headers, body: jsonString);
+
+          // ZIP CHECK
+          http.Response zipResponse = await http
+              .get('https://localhelper-backend.herokuapp.com/api/zips',
+                  headers: headers)
+              .timeout(Duration(seconds: 5));
+
+          var zipJson = jsonDecode(zipResponse.body);
+
+          int _zipID = -1;
+          String _zip = "";
+          bool _found = false;
+          for (int i = 0; i < zipJson.length; i++) {
+            if (zipJson[i]['zip'] == zipController.text.toString()) {
+              _zipID = zipJson[i]['id'];
+              _zip = zipJson[i]['zip'];
+              _found = true;
+              break;
+            }
+          }
+
+          // If the zip was in the database
+          if (_found) {
+            authSettings.zipID = _zipID;
+            authSettings.zip = _zip;
+
+            // Check if the Zip was already added before
+            http.Response zipCheck = await http
+                .get('https://localhelper-backend.herokuapp.com/api/users/me',
+                    headers: headers)
+                .timeout(Duration(seconds: 5));
+
+            var zipCheckJson = jsonDecode(zipCheck.body);
+
+            bool _found = false;
+            for (int i = 0; i < zipCheckJson['zips'].length; i++) {
+              if (zipCheckJson['zips'][i]['zip'] == _zip) {
+                _found = true;
+                break;
+              }
+            }
+
+            // If it wasn't added before add it
+            if (!_found) {
+              Map<String, dynamic> jsonMap = {
+                'userId': authSettings.ownerId,
+                'zipId': _zipID,
+              };
+
+              // Encode
+              String jsonString = json.encode(jsonMap);
+
+              http.Response zipPut = await http
+                  .post(
+                      'https://localhelper-backend.herokuapp.com/api/userZips',
+                      headers: headers,
+                      body: jsonString)
+                  .timeout(Duration(seconds: 5));
+              print(zipPut.statusCode);
+            }
+          }
+
+          // If it wasn't in the database add a new one
+          else {
+            // Put body
+            Map<String, dynamic> jsonMap = {
+              'zip': zipController.text,
+              'name': "NEW",
+            };
+
+            // Encode
+            String jsonString = json.encode(jsonMap);
+
+            // Header
+            Map<String, String> headers = {
+              'authorization': authSettings.token,
+              "Content-Type": "application/json",
+            };
+            String link = 'https://localhelper-backend.herokuapp.com/api/zips';
+            var newZip =
+                await http.post(link, headers: headers, body: jsonString);
+
+            var newJson = jsonDecode(newZip.body);
+
+            authSettings.zipID = newJson['id'];
+            authSettings.zip = newJson['zip'];
+          }
+
+          if (result.statusCode == 200) {
+            // Update info
+            authSettings.updateFirst(jsonMap['first']);
+            authSettings.updateLast(jsonMap['last']);
+            authSettings.updateGender(jsonMap['gender']);
+            authSettings.updatePhone(jsonMap['phone']);
+            authSettings.updateEmail(jsonMap['email']);
+
+            setState(() {
+              _loading = false;
+            });
+            Navigator.pop(context);
+          }
+        } catch (e) {
+          print(e);
+        }
+
+        setState(() {
+          _loading = false;
+        });
+      },
+      child: _loading
+          ? CircularProgressIndicator()
+          : Text(
+              'Save',
+              style: TextStyle(
+                fontSize: 30,
+                fontWeight: FontWeight.bold,
+                color: settings.darkMode ? Colors.white : Colors.black,
+              ),
+            ),
+    );
+  }
+
+  // Save button function
+  Widget deleteButton(Settings settings, AuthSettings authSettings) {
+    return FlatButton(
+      color: Colors.red[300],
+      height: 80,
+      minWidth: double.infinity,
+      onPressed: () async {
+        bool answer = await _showMyDialog();
+
+        if (answer) {
+          try {
+            // Header
+            Map<String, String> headers = {
+              'authorization': authSettings.token,
+              "Content-Type": "application/json",
+            };
+            String link = 'https://localhelper-backend.herokuapp.com/api/users';
+            var result = await http.delete(link, headers: headers);
+            print(result.statusCode);
+            if (result.statusCode == 200) {
+              setState(() {
+                _loading = false;
+              });
+              Navigator.of(context).popUntil((route) => route.isFirst);
+            }
+          } catch (e) {
+            print(e);
+          }
+
+          setState(() {
+            _loading = false;
+          });
+        }
+      },
+      child: _loading
+          ? CircularProgressIndicator()
+          : Text(
+              'Delete Account',
+              style: TextStyle(
+                fontSize: 30,
+                fontWeight: FontWeight.bold,
+                color: settings.darkMode ? Colors.white : Colors.black,
+              ),
+            ),
+    );
+  }
+
+// =============================================================================
+// MAIN ========================================================================
 
   @override
   Widget build(BuildContext context) {
@@ -322,241 +569,6 @@ class _OwnerDoneState extends State<OwnerDone> {
           ],
         ),
       ),
-    );
-  }
-
-  // Save button function
-  Widget saveButton(Settings settings, AuthSettings authSettings) {
-    return FlatButton(
-      height: 60,
-      minWidth: double.infinity,
-      onPressed: () async {
-        if (zipController.text.length < 5) {
-          setState(() {
-            _invalidZip = true;
-          });
-          return false;
-        }
-
-        try {
-          // Put body
-          Map<String, dynamic> jsonMap = {
-            'first': firstNController.text,
-            'last': lastNController.text,
-            'gender': genderController.text,
-            'phone': phoneController.text,
-            'email': emailController.text,
-          };
-
-          // Encode
-          String jsonString = json.encode(jsonMap);
-
-          // Header
-          Map<String, String> headers = {
-            'authorization': authSettings.token,
-            "Content-Type": "application/json",
-          };
-          String link = 'https://localhelper-backend.herokuapp.com/api/users';
-          var result = await http.put(link, headers: headers, body: jsonString);
-
-          // ZIP CHECK
-          http.Response zipResponse = await http
-              .get('https://localhelper-backend.herokuapp.com/api/zips',
-                  headers: headers)
-              .timeout(Duration(seconds: 5));
-
-          var zipJson = jsonDecode(zipResponse.body);
-
-          int _zipID = -1;
-          String _zip = "";
-          bool _found = false;
-          for (int i = 0; i < zipJson.length; i++) {
-            if (zipJson[i]['zip'] == zipController.text.toString()) {
-              _zipID = zipJson[i]['id'];
-              _zip = zipJson[i]['zip'];
-              _found = true;
-              break;
-            }
-          }
-
-          // If the zip was in the database
-          if (_found) {
-            authSettings.zipID = _zipID;
-            authSettings.zip = _zip;
-
-            // Check if the Zip was already added before
-            http.Response zipCheck = await http
-                .get('https://localhelper-backend.herokuapp.com/api/users/me',
-                    headers: headers)
-                .timeout(Duration(seconds: 5));
-
-            var zipCheckJson = jsonDecode(zipCheck.body);
-
-            bool _found = false;
-            for (int i = 0; i < zipCheckJson['zips'].length; i++) {
-              if (zipCheckJson['zips'][i]['zip'] == _zip) {
-                _found = true;
-                break;
-              }
-            }
-
-            // If it wasn't added before add it
-            if (!_found) {
-              Map<String, dynamic> jsonMap = {
-                'userId': authSettings.ownerId,
-                'zipId': _zipID,
-              };
-
-              // Encode
-              String jsonString = json.encode(jsonMap);
-
-              http.Response zipPut = await http
-                  .post(
-                      'https://localhelper-backend.herokuapp.com/api/userZips',
-                      headers: headers,
-                      body: jsonString)
-                  .timeout(Duration(seconds: 5));
-              print(zipPut.statusCode);
-            }
-          }
-
-          // If it wasn't in the database add a new one
-          else {
-// Put body
-            Map<String, dynamic> jsonMap = {
-              'zip': zipController.text,
-              'name': "NEW",
-            };
-
-            // Encode
-            String jsonString = json.encode(jsonMap);
-
-            // Header
-            Map<String, String> headers = {
-              'authorization': authSettings.token,
-              "Content-Type": "application/json",
-            };
-            String link = 'https://localhelper-backend.herokuapp.com/api/zips';
-            var newZip =
-                await http.post(link, headers: headers, body: jsonString);
-
-            var newJson = jsonDecode(newZip.body);
-
-            authSettings.zipID = newJson['id'];
-            authSettings.zip = newJson['zip'];
-          }
-
-          if (result.statusCode == 200) {
-            // Update info
-            authSettings.updateFirst(jsonMap['first']);
-            authSettings.updateLast(jsonMap['last']);
-            authSettings.updateGender(jsonMap['gender']);
-            authSettings.updatePhone(jsonMap['phone']);
-            authSettings.updateEmail(jsonMap['email']);
-
-            setState(() {
-              _loading = false;
-            });
-            Navigator.pop(context);
-          }
-        } catch (e) {
-          print(e);
-        }
-
-        setState(() {
-          _loading = false;
-        });
-      },
-      child: _loading
-          ? CircularProgressIndicator()
-          : Text(
-              'Save',
-              style: TextStyle(
-                fontSize: 30,
-                fontWeight: FontWeight.bold,
-                color: settings.darkMode ? Colors.white : Colors.black,
-              ),
-            ),
-    );
-  }
-
-  // Save button function
-  Widget deleteButton(Settings settings, AuthSettings authSettings) {
-    return FlatButton(
-      color: Colors.red[300],
-      height: 80,
-      minWidth: double.infinity,
-      onPressed: () async {
-        bool answer = await _showMyDialog();
-
-        if (answer) {
-          try {
-            // Header
-            Map<String, String> headers = {
-              'authorization': authSettings.token,
-              "Content-Type": "application/json",
-            };
-            String link = 'https://localhelper-backend.herokuapp.com/api/users';
-            var result = await http.delete(link, headers: headers);
-            print(result.statusCode);
-            if (result.statusCode == 200) {
-              setState(() {
-                _loading = false;
-              });
-              Navigator.of(context).popUntil((route) => route.isFirst);
-            }
-          } catch (e) {
-            print(e);
-          }
-
-          setState(() {
-            _loading = false;
-          });
-        }
-      },
-      child: _loading
-          ? CircularProgressIndicator()
-          : Text(
-              'Delete Account',
-              style: TextStyle(
-                fontSize: 30,
-                fontWeight: FontWeight.bold,
-                color: settings.darkMode ? Colors.white : Colors.black,
-              ),
-            ),
-    );
-  }
-
-  Future<bool> _showMyDialog() async {
-    return showDialog<bool>(
-      context: context,
-      barrierDismissible: false, // user must tap button!
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text('Delete User?'),
-          content: SingleChildScrollView(
-            child: ListBody(
-              children: <Widget>[
-                Text('This action is permanent!'),
-              ],
-            ),
-          ),
-          actions: <Widget>[
-            TextButton(
-              child: Text('No'),
-              onPressed: () {
-                Navigator.of(context).pop(false);
-              },
-            ),
-            TextButton(
-              child: Text('Yes'),
-              onPressed: () {
-                Navigator.of(context).pop(true);
-              },
-            ),
-          ],
-        );
-      },
     );
   }
 }
