+++ frontend/LocalHelper/pubspec.yaml
@@ -24,14 +24,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  firebase_core: ^0.5.0
-  firebase_auth: ^0.18.0+1
-  firebase_database: ^4.1.1
-  lit_firebase_auth: ^0.3.0
   pull_to_refresh: ^1.6.1
   http: ^0.12.2
   provider: ^4.3.2+2
-  geolocator: ^6.0.0+4
 
 dev_dependencies:
   flutter_test:
